Project file
Consider an information system to be used for tracking whether patients admitted to Boston Medical Center for COVID-19 are eligible for discharge based on their health status and where they should be discharged. The general policy is described in the document “COVID19 Clinical Criteria.” We assume that patient information, including tests, measurements, and medications administered, are recorded in our database.